Fix car and cafe shift pages needing perms they should not need
